requirements.txt
# BalPoE-CalibratedLT + MoE-Plugin Pipeline Requirements
# Compatible with Python 3.8

# Core ML dependencies
torch>=1.11.0,<2.0.0
torchvision>=0.12.0,<0.15.0
numpy>=1.19.0,<1.24.0
matplotlib>=3.3.0,<3.7.0
tqdm>=4.64.0
pandas>=1.3.0,<2.0.0
tensorboard>=2.8.0,<2.12.0
Pillow>=8.3.0,<10.0.0
pyyaml>=5.4.0
scikit-learn>=1.0.0,<1.2.0
h5py>=3.1.0,<3.8.0
lightning>=1.6.0,<1.9.0
easydict>=1.9

# Additional dependencies for MoE-Plugin pipeline
scipy>=1.7.0,<1.10.0
seaborn>=0.11.0,<0.12.0
plotly>=5.0.0,<5.15.0

# Development and analysis tools
jupyter>=1.0.0,<1.1.0
ipykernel>=6.0.0,<6.16.0
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0

# Additional utilities
pathlib2>=2.3.0; python_version<"3.4"
typing-extensions>=3.10.0; python_version<"3.8"